rtc: s2mps11: fix invalid type of values
sec_rtc_read, sec_rtc_update functions are required to
use unsigned int value. but, s2mps11 driver used u8 type.
This problem occurs kernel panic due to kernel NULL pointer.

This patch fixes the above problem.

+++ drivers/mfd/sec-core.c
@@ -117,7 +117,8 @@ int sec_rtc_bulk_write(struct sec_pmic_dev *sec_pmic, u8 reg, int count, u8 *buf
 }
 EXPORT_SYMBOL_GPL(sec_rtc_bulk_write);
 
-int sec_rtc_update(struct sec_pmic_dev *sec_pmic, u8 reg, u8 val, u8 mask)
+int sec_rtc_update(struct sec_pmic_dev *sec_pmic, u8 reg, unsigned int val,
+		unsigned int mask)
 {
 	return regmap_update_bits(sec_pmic->rtc_regmap, reg, mask, val);
 }
